benchmarks/cdk/requirements.txt
# AWS CDK dependencies for Thunderduck benchmark infrastructure
aws-cdk-lib>=2.100.0
constructs>=10.0.0
